Use resources for both in and output
